Detect a general protection fault #15
