+++ kernel/src/x64/task/state/segment/descriptor.rs
@@ -23,8 +23,8 @@ impl Descriptor {
     pub fn base_address(&self) -> Option<usize> {
         let higher_base_address: usize = (self.base() as usize) << u32::BITS;
         let lower_descriptor: memory::segment::short::Descriptor = self.lower_descriptor();
-        let lower_descriptor: Option<memory::segment::short::Interface> = (&lower_descriptor).into();
-        lower_descriptor.map(|lower_descriptor| lower_descriptor.base_address() + higher_base_address)
+        let lower_descriptor: Option<memory::segment::short::descriptor::Interface> = (&lower_descriptor).into();
+        lower_descriptor.map(|lower_descriptor| lower_descriptor.base() + higher_base_address)
     }
 
     pub fn lower_descriptor(&self) -> memory::segment::short::Descriptor {
